Moved Personality out of cats.py, changed mentor facet checks to only call the personality of the mentor to remove dependency on Cat.
