+++ scripts/patrol/patrol_outcome.py
@@ -780,16 +780,14 @@ class PatrolOutcome:
 
         for cat in patrol.patrol_cats:
             if Cat.fetch_cat(cat.mentor) in patrol.patrol_cats:
+                mentor = Cat.fetch_cat(cat.mentor)
                 affect_personality = cat.personality.mentor_influence(
-                    Cat.fetch_cat(cat.mentor)
+                    mentor.personality
                 )
-                affect_skills = cat.skills.mentor_influence(Cat.fetch_cat(cat.mentor))
+                affect_skills = cat.skills.mentor_influence(mentor)
                 if affect_personality:
                     History.add_facet_mentor_influence(
-                        cat,
-                        affect_personality[0],
-                        affect_personality[1],
-                        affect_personality[2],
+                        cat, mentor.ID, affect_personality[0], affect_personality[1]
                     )
                     print(str(cat.name), affect_personality)
                 if affect_skills:
